use UTF-8 for resources in Oomph setup
According to the POMs the sources are all UTF-8, but the Oomph setup
does not set an encoding. On Windows the default encoding is not UTF-8,
therefore this easily leads to confusion in mixed OS environments.

The additional compound task was created by the preference capture
mechanism. I suggest to leave it in, as it is more easy to capture
additional preferences into the setup in the future.

Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>